fixup! gvfs-helper: create tool to fetch objects using the GVFS Protocol
It would appear that libcurl 7.68.0-1ubuntu2.25 sets the
`CURLOPT_FAILONERROR` option to true by default, but we need to parse
the headers even in the error case.

This results in a slight change of behavior: When authentication fails,
the connection is now kept open. Testament to this behavioral change is
t5799.24(HTTP GET Auth on Origin Server) which wanted to verify that
there are two connections when authentication is required, but now there
is only one (the rest of the test case verifies that fetching the loose
object worked, i.e. that the authentication succeeded).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>