fixup! xdiff: make fields of xrecord_t Rust friendly
To make `xdl_classify_record()` work, the `ha` attributes of `xrecord_t`
and of `s_xdlclass` _must_ have the same range. Otherwise the function
won't be able to recognize previously-classified records correctly when
the `ha` recorded in the `xrecord_t` is so wide that it won't fit into
the `s_xdlclass`' attribute and therefore they won't match when they
need to match.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>